+++ src/yamlServerInit.ts
@@ -91,8 +91,12 @@ export class YAMLServerInit {
         },
         documentRangeFormattingProvider: false,
         documentLinkProvider: {},
-        foldingRangeProvider: true,
+        // disabled until we not get parser which parse comments as separate nodes
+        foldingRangeProvider: false,
         codeActionProvider: true,
+        codeLensProvider: {
+          resolveProvider: false,
+        },
         executeCommandProvider: {
           commands: Object.keys(YamlCommands).map((k) => YamlCommands[k]),
         },